==> notebooks/deliveries/1_clean_traffic_sensors.ipynb <==
import os
import pandas as pd

# Define the root directory
root_dir = '../../data/raw/traffic_sensors_locations'

# Create an auxiliary DataFrame to store unique latitud and longitud
aux_df = pd.DataFrame(columns=['latitud', 'longitud'])

# Iterate through all year folders
for year in range(2018, 2026):
    year_folder = os.path.join(root_dir, str(year))
    
    # Check if the year folder exists
    if os.path.isdir(year_folder):
        print(f"Processing folder: {year_folder}")
        
        # Iterate through files in the year folder
        for file_name in os.listdir(year_folder):
            if file_name.endswith('.xlsx'):  # Ensure it's a CSV file
                
                file_path = os.path.join(year_folder, file_name)
                
                try:
                    df = pd.read_excel(file_path)
                    
                    # Remove duplicated values by looking at both "latitud" and "longitud"
                    df_unique = df[['latitud', 'longitud','id']].drop_duplicates()
                    
                    # Round both columns to 4 decimal places
                    #df_unique['latitud'] = df_unique['latitud'].round(4)
                    #df_unique['longitud'] = df_unique['longitud'].round(4)
                    df_unique['id'] = df_unique['id']
                    
                    # Append to the auxiliary DataFrame
                    aux_df = pd.concat([aux_df, df_unique], ignore_index=True)
                
                    print(f"Processed file: {file_name}")
                except Exception as e:
                    print(f"Error processing file {file_name}: {e}")

df_unique = aux_df.drop_duplicates(subset=['id'], keep='first')

df_unique['id'] = df_unique['id'].astype(int)
df_unique.rename(columns = {'id': 'id_trafico'}, inplace = True)

df_unique.to_excel("1_all_traffic_sensors.xlsx", index = False)
