==> notebooks/deliveries/2_prepare_air_data.ipynb <==
import pandas as pd
import os

# # Iterate over all folders inside ../data/data_quality/ and inside each folder, rename the file by the same
# # name but only with the first 3 characters of the file name.
# # only do it for the .csv files
# # don't do it for the hidden files
# # delete all the other files that are not .csv

# for folder in os.listdir('../../data/air_quality/'):
#     if not folder.startswith('.'):
#         for file in os.listdir(f'../../data/air_quality/{folder}'):
#             if file.endswith('.csv') and not file.startswith('.'):
#                 os.rename(f'../../data/air_quality/{folder}/{file}', f'../../data/air_quality/{folder}/{file[:3]}.csv')
#             else:
#                 os.remove(f'../../data/air_quality/{folder}/{file}')

# # Iterate over all folders and create one big dataframe with all the data
# df = pd.DataFrame()
# for folder in os.listdir('../../data/air_quality/'):
#     if not folder.startswith('.'):
#         for file in os.listdir(f'../../data/air_quality/{folder}'):
#             df = pd.concat([df, pd.read_csv(f'../../data/air_quality/{folder}/{file}', delimiter=';')])

# Import tqdm
from tqdm import tqdm
import os
import pandas as pd

# Iterate over all folders and create one big dataframe with all the data
df = pd.DataFrame()
for folder in tqdm(os.listdir('../../data/raw/air/'), desc='Processing folders'):
    if folder == 'Anio2013' or folder == 'Anio2014' or folder == 'Anio2015' or folder == 'Anio2016' or folder == 'Anio2017':
        continue
    if not folder.startswith('.'):
        for file in tqdm(os.listdir(f'../../data/raw/air/{folder}'), desc=f'Processing files in {folder}', leave=False):
            if file.endswith('.csv'):  # Only process .csv files
                df = pd.concat([df, pd.read_csv(f'../../data/raw/air/{folder}/{file}', delimiter=';')])

df.to_parquet('2_all_data_air_quality_from_2018.parquet', index=False)
