==> notebooks/deliveries/6_generate_air_data.ipynb <==
import pandas as pd

pd.set_option('display.max_columns', None)
pd.set_option('display.width', None)

## Imports

df = pd.read_parquet("../data_deliveries/3_df_air_quality_and_locations_from_2018.parquet")

df = df[['CODIGO', 'FECHA', 'NO2_VALUE', 'LONGITUD', 'LATITUD']].rename(columns=str.lower)
df = df.rename(columns = {'codigo':'id_no2'})

df.to_csv("6_df_air_data_and_locations_reduced.csv", index = False)
